Move button_author_validation() from pagination.py to component_helper.py
